To run zb test: `python test_runner.py ./test_out --test pp_zb`

TODO:
- make dw runner actually update the weight gradients



[ghstack-poisoned]

+++ test_runner.py
@@ -53,14 +53,29 @@ def build_test_list():
                     "--checkpoint.enable_checkpoint",
                     "--experimental.pipeline_parallel_degree 4",
                     "--experimental.pipeline_parallel_split_points layers.1,layers.2,layers.3,layers.4,layers.5,layers.6,layers.7",
+                    "--experimental.pipeline_parallel_schedule flexible_interleaved_1f1b",
+                    "--model.norm_type rmsnorm",  # fused_rmsnorm throws cuda context error with pp
+                ],
+            ],
+            "PP looped flexible 1f1b test",
+            "pp_looped_flexible_1f1b",
+            requires_seed_checkpoint=True,
+            ngpu=4,
+        ),
+        OverrideDefinitions(
+            [
+                [
+                    "--checkpoint.enable_checkpoint",
+                    "--experimental.pipeline_parallel_degree 4",
+                    "--experimental.pipeline_parallel_split_points layers.1,layers.2,layers.3,layers.4,layers.5,layers.6,layers.7",
                     "--experimental.pipeline_parallel_microbatches 8",
-                    "--experimental.pipeline_parallel_schedule zb",
+                    "--experimental.pipeline_parallel_schedule zb_v",
                     "--training.data_parallel_degree 1",
                     "--model.norm_type rmsnorm",  # fused_rmsnorm crashes with PP
                 ],
             ],
-            "PP fake ZB test",
-            "pp_zb",
+            "PP fake ZB test with v shaped stages",
+            "pp_zb_v",
             requires_seed_checkpoint=True,
             ngpu=4,
         ),
@@ -170,6 +185,17 @@ def build_test_list():
             ],
             "1D compile",
             "1d_compile",
+        ),
+        OverrideDefinitions(
+            [
+                [
+                    "--training.compile",
+                    "--activation_checkpoint.mode selective",
+                    "--activation_checkpoint.selective_ac_option op",
+                ],
+            ],
+            "1D compile with selective op AC",
+            "1d_compile_sac_op",
         ),
         OverrideDefinitions(
             [
@@ -332,6 +358,8 @@ def run_test(test_flavor: OverrideDefinitions, full_path: str, output_dir: str):
 
     for override_arg in test_flavor.override_args:
         cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_llama_train.sh"
+        if test_name == "fsdp2_mem_tracker":
+            cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_memory_estimation.sh"
         cmd += " " + dump_folder_arg
         cmd += " " + model_flavor_arg
         if override_arg:
@@ -388,8 +416,11 @@ def main():
     if not os.path.exists(args.output_dir):
         os.makedirs(args.output_dir)
     if os.listdir(args.output_dir):
-        # raise RuntimeError("Please provide an empty output directory.")
-        shutil.rmtree(args.output_dir)
+        if args.output_dir == "./test_out":
+            shutil.rmtree(args.output_dir)
+            os.makedirs(args.output_dir)  # Recreate the directory after removing it
+        else:
+            raise RuntimeError("Please provide an empty output directory.")
     run_tests(args)
